requirements.txt
Django==1.5
Markdown==2.3.1
django-filter==0.6
django-reversion==1.7.1
djangorestframework==2.3.6
wsgiref==0.1.2
